backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==1.2.0
pymcprotocol==0.3.6
psycopg2-binary==2.9.7
python-dateutil==2.8.2
gunicorn==21.2.0